refactor: split `init.py` into multiple modules
BREAKING CHANGE: delete `init.py` and split its functions into modules in `.../simulat/core/init/`
